Rename exporters/metric/prometheus
