+++ example/prom-collector/go.mod
@@ -4,15 +4,15 @@ go 1.15
 
 replace (
 	go.opentelemetry.io/otel => ../..
-	go.opentelemetry.io/otel/exporters/metric/prometheus => ../../exporters/metric/prometheus
 	go.opentelemetry.io/otel/exporters/otlp => ../../exporters/otlp
+	go.opentelemetry.io/otel/exporters/prometheus => ../../exporters/prometheus
 	go.opentelemetry.io/otel/sdk => ../../sdk
 )
 
 require (
 	go.opentelemetry.io/otel v0.20.0
-	go.opentelemetry.io/otel/exporters/metric/prometheus v0.20.0
 	go.opentelemetry.io/otel/exporters/otlp v0.20.0
+	go.opentelemetry.io/otel/exporters/prometheus v0.20.0
 	go.opentelemetry.io/otel/metric v0.20.0
 	go.opentelemetry.io/otel/sdk v0.20.0
 	go.opentelemetry.io/otel/sdk/metric v0.20.0
