+++ exporters/otlp/go.mod
@@ -42,7 +42,7 @@ replace go.opentelemetry.io/otel/example/prometheus => ../../example/prometheus
 
 replace go.opentelemetry.io/otel/example/zipkin => ../../example/zipkin
 
-replace go.opentelemetry.io/otel/exporters/metric/prometheus => ../metric/prometheus
+replace go.opentelemetry.io/otel/exporters/prometheus => ../prometheus
 
 replace go.opentelemetry.io/otel/exporters/otlp => ./
