+++ exporters/otlp/otlptrace/otlptracegrpc/go.mod
@@ -45,7 +45,7 @@ replace go.opentelemetry.io/otel/example/prometheus => ../../../../example/prome
 
 replace go.opentelemetry.io/otel/example/zipkin => ../../../../example/zipkin
 
-replace go.opentelemetry.io/otel/exporters/metric/prometheus => ../../../metric/prometheus
+replace go.opentelemetry.io/otel/exporters/prometheus => ../../../prometheus
 
 replace go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc => ./
