+++ exporters/prometheus/README.md
@@ -5,5 +5,5 @@ OpenTelemetry Prometheus exporter
 ## Installation
 
 ```
-go get -u go.opentelemetry.io/otel/exporters/metric/prometheus
+go get -u go.opentelemetry.io/otel/exporters/prometheus
 ```
