+++ exporters/prometheus/go.mod
@@ -1,18 +1,18 @@
-module go.opentelemetry.io/otel/internal/metric
+module go.opentelemetry.io/otel/exporters/prometheus
 
 go 1.15
 
 require (
+	github.com/prometheus/client_golang v1.11.0
 	github.com/stretchr/testify v1.7.0
 	go.opentelemetry.io/otel v0.20.0
 	go.opentelemetry.io/otel/metric v0.20.0
+	go.opentelemetry.io/otel/sdk v0.20.0
+	go.opentelemetry.io/otel/sdk/export/metric v0.20.0
+	go.opentelemetry.io/otel/sdk/metric v0.20.0
 )
 
 replace go.opentelemetry.io/otel => ../..
-
-replace go.opentelemetry.io/otel/metric => ../../metric
-
-replace go.opentelemetry.io/otel/internal/metric => ./
 
 replace go.opentelemetry.io/otel/bridge/opencensus => ../../bridge/opencensus
 
@@ -34,23 +34,27 @@ replace go.opentelemetry.io/otel/example/prometheus => ../../example/prometheus
 
 replace go.opentelemetry.io/otel/example/zipkin => ../../example/zipkin
 
-replace go.opentelemetry.io/otel/exporters/prometheus => ../../exporters/prometheus
+replace go.opentelemetry.io/otel/exporters/otlp => ../otlp
 
-replace go.opentelemetry.io/otel/exporters/otlp => ../../exporters/otlp
+replace go.opentelemetry.io/otel/exporters/otlp/otlptrace => ../otlp/otlptrace
 
-replace go.opentelemetry.io/otel/exporters/otlp/otlptrace => ../../exporters/otlp/otlptrace
+replace go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc => ../otlp/otlptrace/otlptracegrpc
 
-replace go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc => ../../exporters/otlp/otlptrace/otlptracegrpc
+replace go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp => ../otlp/otlptrace/otlptracehttp
 
-replace go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp => ../../exporters/otlp/otlptrace/otlptracehttp
+replace go.opentelemetry.io/otel/exporters/prometheus => ./
 
-replace go.opentelemetry.io/otel/exporters/stdout => ../../exporters/stdout
+replace go.opentelemetry.io/otel/exporters/stdout => ../stdout
 
-replace go.opentelemetry.io/otel/exporters/trace/jaeger => ../../exporters/trace/jaeger
+replace go.opentelemetry.io/otel/exporters/trace/jaeger => ../trace/jaeger
 
-replace go.opentelemetry.io/otel/exporters/trace/zipkin => ../../exporters/trace/zipkin
+replace go.opentelemetry.io/otel/exporters/trace/zipkin => ../trace/zipkin
 
-replace go.opentelemetry.io/otel/internal/tools => ../tools
+replace go.opentelemetry.io/otel/internal/metric => ../../internal/metric
+
+replace go.opentelemetry.io/otel/internal/tools => ../../internal/tools
+
+replace go.opentelemetry.io/otel/metric => ../../metric
 
 replace go.opentelemetry.io/otel/oteltest => ../../oteltest
