+++ exporters/prometheus/sanitize.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package prometheus // import "go.opentelemetry.io/otel/exporters/metric/prometheus"
+package prometheus // import "go.opentelemetry.io/otel/exporters/prometheus"
 
 import (
 	"strings"
