+++ exporters/stdout/go.mod
@@ -36,7 +36,7 @@ replace go.opentelemetry.io/otel/example/prometheus => ../../example/prometheus
 
 replace go.opentelemetry.io/otel/example/zipkin => ../../example/zipkin
 
-replace go.opentelemetry.io/otel/exporters/metric/prometheus => ../metric/prometheus
+replace go.opentelemetry.io/otel/exporters/prometheus => ../prometheus
 
 replace go.opentelemetry.io/otel/exporters/otlp => ../otlp
