+++ sdk/export/metric/go.mod
@@ -22,7 +22,7 @@ replace go.opentelemetry.io/otel/example/prometheus => ../../../example/promethe
 
 replace go.opentelemetry.io/otel/example/zipkin => ../../../example/zipkin
 
-replace go.opentelemetry.io/otel/exporters/metric/prometheus => ../../../exporters/metric/prometheus
+replace go.opentelemetry.io/otel/exporters/prometheus => ../../../exporters/prometheus
 
 replace go.opentelemetry.io/otel/exporters/otlp => ../../../exporters/otlp
