+++ sdk/go.mod
@@ -30,7 +30,7 @@ replace go.opentelemetry.io/otel/example/prometheus => ../example/prometheus
 
 replace go.opentelemetry.io/otel/example/zipkin => ../example/zipkin
 
-replace go.opentelemetry.io/otel/exporters/metric/prometheus => ../exporters/metric/prometheus
+replace go.opentelemetry.io/otel/exporters/prometheus => ../exporters/prometheus
 
 replace go.opentelemetry.io/otel/exporters/otlp => ../exporters/otlp
